Bugfix: reverse() also pushed into _dfs()
This is a further fix on #15a4eadfdea8

in that commit
SortedSet(reversed(...dfs

was replaced by
reversed(dfs(...

where dfs contained the SortedSet.

This broke the order of the dependencies.

I've now restored this by pushing reversed into _dfs. Rather than
calling reversed() all the time I've changed the appending of results
into prepending.

This is all in untested territory. I.e. tests run but they
also did when stuff was broken in the previous commits.